aistudio/MathorCup2025/requirements.txt
open3d==0.17.0
paddlesci
meshio
